log_to_story/requirements.txt
fastapi>=0.95.0
uvicorn[standard]
python-multipart
aiofiles
python-dateutil
sentence-transformers
faiss-cpu
sqlalchemy
httpx
